requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
psycopg2-binary==2.9.5